requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mangum==0.17.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
playwright==1.40.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pandas==2.1.3
reportlab==4.0.7
openpyxl==3.1.2
fastapi-cors==0.0.6

